Make use of sparsity in proving rational sum for f
While the measurements have high variance, in theory this should increase performance by up to 50% (exact number tbd) in compute cubic and bind. In practice it seems to be about 40%-50% faster. But again, this is kind of a curated benchmark for it, so take the numbers with a grain of salt.